here we go

+++ GameRoullete/GameDataBase/Models/GameRating.cs
@@ -8,10 +8,10 @@ namespace Game_Collector.Models
 {
    public class GameRating
     {
-        public decimal game_Total_Rating { get; set; }
-        public int game_Total_Rating_Count { get; set; }
-        public int game_popularity { get; set; }
-        public int game_Hype { get; set; }
+        public double game_Total_Rating { get; set; }
+        public double game_Total_Rating_Count { get; set; }
+        public double   game_popularity { get; set; }
+        public double game_Hype { get; set; }
         public int game_id { get; set; } //PK/FK
     }
 }
